requirements.txt
flask
gunicorn
itsdangerous==2.1.2
Jinja2==3.1.2
MarkupSafe==2.1.2
python-dotenv
Werkzeug==2.2.3
langchain_openai 
langchain_core
python-dotenv
langchain_community
pypdf
groq
langchain-groq
langchainhub
langchain
pinecone
pygame
pinecone
translate
# pyttsx3
# SpeechRecognition
# PyAudio
# eSpeak-ng

